requirements.txt
numpy==1.16.0
Pillow==5.4.1
opencv-python==4.0.0.21
